BF+TST: Fix issues with weird chars in filenames, add tests
Fixes issues with newlines, tabs, and invalid unicode in paths
that are being backed up. Adds sorely lacking tests for the
jobs.py module and the CLI.

Added '--max-jobs' option to CLI for 'daemon' command so that
it will exit after running that many jobs. Useful for testing,
but also in case we want to backup the db/indices after running
some jobs.