ENH: Loop and recheck drive status if it is out of sync
Apparently the 'mtx' command can just return success even when it
completely fails, even though it isn't even done trying. At some
later point it might even succeed at what it was originally supposed
to do...

To work around this stupidity and not always have to wait 5+ minutes
on every tape load, we dynamically recheck the drive status for a while
when it looks like things are out of sync. This way if the task does
eventually complete we can avoid bailing out.